# packages/backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "WallStr.chat",
        "summary": "WallStr.chat OpenAPI schema",
        "version": "0.1.0"
    },
    "paths": {
...
        "/chats/{slug}": {
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "Get Chat",
                "operationId": "2::get_chat",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Slug"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPUnauthorizedError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chats/{slug}/messages": {
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "Get Chat Messages",
                "operationId": "2::get_chat_messages",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Slug"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Cursor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paginated_ChatMessage_"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPUnauthorizedError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "chat"
                ],
                "summary": "Send Chat Message",
                "operationId": "2::send_chat_message",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Slug"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatMessage"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPUnauthorizedError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Chat not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chats": {
            "post": {
                "tags": [
                    "chat"
                ],
                "summary": "Create Chat",
                "operationId": "2::create_chat",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPUnauthorizedError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "List Chats",
                "operationId": "2::list_chats",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Cursor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paginated_Chat_"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPUnauthorizedError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Chat": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "slug": {
                        "type": "string",
                        "title": "Slug"
                    },
                    "title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Title"
                    },
                    "messages": {
                        "$ref": "#/components/schemas/Paginated_ChatMessage_"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "slug",
                    "title",
                    "messages"
                ],
                "title": "Chat"
            },
            "ChatMessage": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "role": {
                        "$ref": "#/components/schemas/ChatMessageRole"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "message_type": {
                        "$ref": "#/components/schemas/ChatMessageType"
                    },
                    "documents": {
                        "items": {
                            "$ref": "#/components/schemas/Document"
                        },
                        "type": "array",
                        "title": "Documents"
                    },
                    "pending_documents": {
                        "items": {
                            "$ref": "#/components/schemas/PendingDocument"
                        },
                        "type": "array",
                        "title": "Pending Documents",
                        "default": []
                    },
                    "memo": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Memo"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "role",
                    "content",
                    "message_type",
                    "documents",
                    "created_at"
                ],
                "title": "ChatMessage"
            },
            "ChatMessageRole": {
                "type": "string",
                "enum": [
                    "user",
                    "assistant"
                ],
                "title": "ChatMessageRole"
            },
            "ChatMessageType": {
                "type": "string",
                "enum": [
                    "text",
                    "memo"
                ],
                "title": "ChatMessageType"
            },
...
            "Document": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "status": {
                        "$ref": "#/components/schemas/DocumentStatus"
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error"
                    },
                    "errored_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Errored At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename",
                    "status"
                ],
                "title": "Document"
            },
            "DocumentPayload": {
                "properties": {
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "doc_type": {
                        "$ref": "#/components/schemas/DocumentType"
                    }
                },
                "type": "object",
                "required": [
                    "filename",
                    "doc_type"
                ],
                "title": "DocumentPayload"
            },
...
            "DocumentStatus": {
                "type": "string",
                "enum": [
                    "uploading",
                    "uploaded",
                    "processing",
                    "ready"
                ],
                "title": "DocumentStatus"
            },
            "DocumentType": {
                "type": "string",
                "enum": [
                    "pdf",
                    "doc",
                    "docx",
                    "xls",
                    "xlsx"
                ],
                "title": "DocumentType"
            },
            "HTTPUnauthorizedError": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "HTTPUnauthorizedError"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Memo": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "user_prompt": {
                        "type": "string",
                        "title": "User Prompt"
                    },
                    "sections": {
                        "items": {
                            "$ref": "#/components/schemas/MemoSection"
                        },
                        "type": "array",
                        "title": "Sections"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "user_prompt",
                    "sections"
                ],
                "title": "Memo"
            },
            "MemoSection": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "group": {
                        "type": "string",
                        "title": "Group"
                    },
                    "aspect": {
                        "type": "string",
                        "title": "Aspect"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "index": {
                        "type": "integer",
                        "title": "Index"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "group",
                    "aspect",
                    "content",
                    "index"
                ],
                "title": "MemoSection"
            },
            "MessageRequest": {
                "properties": {
                    "message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Message"
                    },
                    "documents": {
                        "items": {
                            "$ref": "#/components/schemas/DocumentPayload"
                        },
                        "type": "array",
                        "title": "Documents"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "documents"
                ],
                "title": "MessageRequest"
            },
            "Paginated_ChatMessage_": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ChatMessage"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "cursor": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cursor"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "cursor"
                ],
                "title": "Paginated[ChatMessage]"
            },
            "Paginated_Chat_": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/Chat"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "cursor": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cursor"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "cursor"
                ],
                "title": "Paginated[Chat]"
            },
            "PendingDocument": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "presigned_url": {
                        "type": "string",
                        "title": "Presigned Url"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename",
                    "presigned_url"
                ],
                "title": "PendingDocument"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}